requirements.txt
discord.py[voice]==2.3.2
python-dotenv==1.0.0
yt-dlp==2023.12.30
PyNaCl==1.5.0
elevenlabs==1.8.1
requests==2.31.0